Fix 3-port AND gate not working on AVR devices
MatIEC uses macros to define arithmetic expressions. Since arithmetic expressions can have any number of arguments, it uses va-args to populate the args list. But there is a catch: va-args does not support any element smaller than the int type. For most platforms, a int is 32 bit wide. Anything smaller than 32 bits is padded to fit 32 bits. This works great everywhere, except on AVR. For some *weird* reason AVR considers int to be 16 bit (instead of 8 since it is a 8 bit CPU). The previous code worked with two gate AND because the first argument is read correctly. But for more than 2 gates, once it tries to catch the third argument, it jumps 32 bits instead of 16 and crashes. So on AVR, va-args hack must be aligned to 16 bit, not 32.