Fixed crashes while compiling some programs
The FactorizePaths() function sorts the factorized_paths array before returning it. On python 2, the list.sort() function allows sorting different data-types. On python 3, this causes an exception. The fix brings a new sorting function that has a similar behavior than Python 2 list.sort(), and uses that instead of the native list.sort() call.